+++ Records/CachedLeve.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Linq;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using Lumina.Excel.GeneratedSheets;
@@ -19,10 +20,15 @@ public record CachedLeve
 
     private Leve? leve = null;
     private string? name = null;
+    private string? nameEn = null;
     private string? levemeteName = null;
-    private string? typeName = null;
+    private uint? townId = null;
     private string? townName = null;
     private RequiredItem[]? requiredItems = null;
+    private CachedLevel? levelLevemete = null;
+    private CachedLeveAssignmentType? leveAssignmentType = null;
+    private int? leveVfxIcon = null;
+    private int? leveVfxFrameIcon = null;
 
     public uint LeveId { get; private set; }
 
@@ -30,15 +36,10 @@ public record CachedLeve
         => leve ??= Service.Data.GetExcelSheet<Leve>()?.GetRow(LeveId);
 
     public string Name
-    {
-        get
-        {
-            if (string.IsNullOrEmpty(name) && Leve != null)
-                name = Leve.Name.ClearString();
+        => name ??= Leve?.Name.ClearString() ?? $"<{LeveId}>";
 
-            return name ?? $"<{LeveId}>";
-        }
-    }
+    public string NameEn
+        => nameEn ??= Service.Data.GetExcelSheet<Leve>(Dalamud.ClientLanguage.English)?.GetRow(LeveId)?.Name.ClearString() ?? $"<{LeveId}>";
 
     public ushort ClassJobLevel
         => Leve?.ClassJobLevel ?? 0;
@@ -57,27 +58,20 @@ public record CachedLeve
         }
     }
 
-    public string TypeName
-    {
-        get
-        {
-            if (string.IsNullOrEmpty(typeName) && Leve != null)
-                typeName = Service.Data.GetExcelSheet<LeveAssignmentType>()?.GetRow((uint)Leve.Unknown4)?.Name.ClearString();
+    public CachedLeveAssignmentType? LeveAssignmentType
+        => leveAssignmentType ??= leve != null ? LeveAssignmentTypeCache.Get((uint)Leve!.Unknown4) : null; // leve.LeveAssignmentType.Row seems to have moved
 
-            return typeName ?? "";
-        }
-    }
+    public int TypeIcon
+        => LeveAssignmentType?.Icon ?? 0;
+
+    public string TypeName
+        => LeveAssignmentType?.Name ?? "";
+
+    public uint TownId
+        => townId ??= Leve?.Town.Row ?? 0;
 
     public string TownName
-    {
-        get
-        {
-            if (string.IsNullOrEmpty(townName) && Leve != null)
-                townName = Leve.Town.Value?.Name.ClearString();
-
-            return townName ?? "???";
-        }
-    }
+        => townName ??= Leve?.Town.Value?.Name.ClearString() ?? "???";
 
     public bool TownLocked
         => Leve != null && (Leve.RowId == 546 || Leve.RowId == 556 || Leve.RowId == 566);
@@ -89,10 +83,10 @@ public record CachedLeve
         => Service.GameFunctions.IsLevequestAccepted(LeveId);
 
     public bool IsCraftLeve
-        => Leve?.LeveAssignmentType.Row is >= 5 and <= 12;
+        => LeveAssignmentType?.RowId is >= 5 and <= 12;
 
     public bool IsGatherLeve
-        => Leve?.LeveAssignmentType.Row is >= 2 and <= 4;
+        => LeveAssignmentType?.RowId is >= 2 and <= 4;
 
     public RequiredItem[]? RequiredItems
     {
@@ -107,10 +101,33 @@ public record CachedLeve
 
             requiredItems = craftLeve.UnkData3
                 .Where(item => item.Item != 0 && item.ItemCount != 0)
-                .Select(item => new RequiredItem(ItemCache.Get((uint)item.Item), item.ItemCount))
+                .Aggregate(
+                    new Dictionary<int, RequiredItem>(),
+                    (dict, entry) =>
+                    {
+                        if (!dict.TryGetValue(entry.Item, out var reqItem))
+                        {
+                            reqItem = new RequiredItem(ItemCache.Get((uint)entry.Item), 0);
+                            dict.Add(entry.Item, reqItem);
+                        }
+
+                        reqItem.Amount += entry.ItemCount;
+
+                        return dict;
+                    })
+                .Values
                 .ToArray();
 
             return requiredItems;
         }
     }
+
+    public CachedLevel? LevelLevemete
+        => levelLevemete ??= Leve != null ? LevelCache.Get(Leve.LevelLevemete.Row) : null;
+
+    public int LeveVfxIcon
+        => leveVfxIcon ??= Leve?.LeveVfx.Value?.Icon ?? 0;
+
+    public int LeveVfxFrameIcon
+        => leveVfxFrameIcon ??= Leve?.LeveVfxFrame.Value?.Icon ?? 0;
 }