+++ Records/CachedLeveAssignmentType.cs
@@ -1,0 +1,26 @@
+using Lumina.Excel.GeneratedSheets;
+
+namespace LeveHelper;
+
+public record CachedLeveAssignmentType
+{
+    public CachedLeveAssignmentType(uint rowId)
+    {
+        RowId = rowId;
+    }
+
+    private LeveAssignmentType? leveAssignmentType { get; set; } = null;
+    private string? name { get; set; } = null;
+    private int? icon { get; set; } = null;
+
+    public uint RowId { get; }
+
+    public LeveAssignmentType? LeveAssignmentType
+        => leveAssignmentType ??= Service.Data.GetExcelSheet<LeveAssignmentType>()?.GetRow(RowId);
+
+    public string Name
+        => name ??= LeveAssignmentType?.Name.ClearString() ?? "";
+
+    public int Icon
+        => icon ??= LeveAssignmentType?.Icon ?? 0;
+}