+++ Utils/GameFunctions.cs
@@ -134,9 +134,13 @@ public unsafe class GameFunctions
     public readonly AgentFishGuide_OpenForItemIdDelegate AgentFishGuide_OpenForItemId = null!;
     public delegate byte* AgentFishGuide_OpenForItemIdDelegate(nint agentFishGuide, uint itemId, bool isSpearfishing);
 
+    [Signature("E8 ?? ?? ?? ?? 48 8B 74 24 ?? 48 8B 7C 24 ?? 48 83 C4 30 5B C3 48 8B CB")]
+    public readonly AgentJournal_OpenForQuestDelegate AgentJournal_OpenForQuest = null!;
+    public delegate byte* AgentJournal_OpenForQuestDelegate(nint agentJournal, int id, int type, ushort a4 = 0, bool a5 = true); // type: 1 = Quest, 2 = Levequest
+
     [Signature("E8 ?? ?? ?? ?? EB 1D 83 F8 0D")]
     private readonly ItemFinderModuleSearchForItemDelegate ItemFinderModule_SearchForItem = null!;
-    private delegate void* ItemFinderModuleSearchForItemDelegate(void* module, uint itemId, bool isHQ = false);
+    private delegate void* ItemFinderModuleSearchForItemDelegate(void* module, uint itemId, bool includeHQ = true);
     public void SearchForItem(uint itemId) => ItemFinderModule_SearchForItem(Framework.Instance()->GetUiModule()->GetItemFinderModule(), itemId);
 
     [Signature("E9 ?? ?? ?? ?? 48 8D 47 30")]